Refactor, extract DTOs into Models/Stats to adhere to SRP
