+++ CaloriesTracker/Models/Stats/CalorieTrendPoint.cs
@@ -1,0 +1,8 @@
+﻿namespace CaloriesTracker.Models.Stats
+{
+    public class CalorieTrendPoint
+    {
+        public DateTime Date { get; set; }
+        public decimal Calories { get; set; }
+    }
+}
