+++ CaloriesTracker/Models/Stats/MacroNutrients.cs
@@ -1,0 +1,9 @@
+﻿namespace CaloriesTracker.Models.Stats
+{
+    public class MacroNutrients
+    {
+        public decimal Protein { get; set; }
+        public decimal Fat { get; set; }
+        public decimal Carbs { get; set; }
+    }
+}
