+++ CaloriesTracker/Views/Home/Index.cshtml
@@ -1,5 +1,5 @@
-@using CaloriesTracker.Services
-@model CalorieService.DailySummary
+@using CaloriesTracker.Models.Stats
+@model DailySummary
 
 @{
     ViewData["Title"] = "Dashboard";
